# fern/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EDON Governance API",
    "version": "1.0.1",
    "description": "Real-time governance for AI agents. Every action evaluated, every decision auditable.",
    "x-fern-sdk-group-name": "edon"
  },
  "servers": [
    {
      "url": "https://edon-gateway-prod.fly.dev",
      "description": "Production"
    }
  ],
  "security": [],
  "paths": {
...
    "/settings/kill-switch": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "Get Kill Switch",
        "description": "Return the current kill switch state for this tenant.\n\nactive=true means ALL agent actions are being blocked right now.",
        "operationId": "get_kill_switch_settings_kill_switch_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "settings"
        ],
        "summary": "Activate",
        "description": "Activate the kill switch \u2014 immediately halt ALL AI agent actions.\n\nEvery /v1/action call for this tenant will return BLOCK until the kill\nswitch is deactivated. The real governance verdict is still computed and\naudit-logged, but the response is always BLOCK.\n\nUse this when:\n- A security incident is suspected\n- An agent is behaving unexpectedly\n- You need an immediate pause before a policy update\n\nThis action is audit-logged with timestamp and reason.",
        "operationId": "activate_settings_kill_switch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KillSwitchActivateBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "settings"
        ],
        "summary": "Deactivate",
        "description": "Deactivate the kill switch \u2014 resume normal governance.\n\nNormal policy evaluation resumes immediately. This action is audit-logged.",
        "operationId": "deactivate_settings_kill_switch_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KillSwitchDeactivateBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "KillSwitchActivateBody": {
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Reason",
            "description": "Why you are halting all AI agents for this tenant"
          },
          "activated_by": {
            "type": "string",
            "title": "Activated By",
            "description": "Identity of the person activating the kill switch",
            "default": "api"
          }
        },
        "type": "object",
        "required": [
          "reason"
        ],
        "title": "KillSwitchActivateBody"
      },
      "KillSwitchDeactivateBody": {
        "properties": {
          "deactivated_by": {
            "type": "string",
            "title": "Deactivated By",
            "default": "api"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "title": "KillSwitchDeactivateBody"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {}
  },
...
}